add jshint loader and config as well as watch to webpack
