+++ package.json
@@ -21,28 +21,37 @@
     "test": "grunt assemble"
   },
   "dependencies": {
+    "6to5-loader": "^2.0.0",
     "assemble": "0.4.41",
+    "banner-footer-webpack-plugin": "git+ssh://git@github.com:dtothefp/banner-footer-webpack-plugin.git",
     "bower": "1.3.8",
+    "exports-loader": "0.6.2",
     "grunt": "^0.4.1",
     "grunt-autoprefixer": "1.0.0",
     "grunt-config": "0.2.1",
     "grunt-contrib-clean": "0.5.0",
     "grunt-contrib-concat": "0.4.0",
     "grunt-contrib-copy": "0.5.0",
-    "grunt-contrib-handlebars": "0.8.0",
     "grunt-contrib-imagemin": "0.7.1",
     "grunt-contrib-jshint": "0.10.0",
-    "grunt-contrib-uglify": "0.5.0",
     "grunt-filerev": "0.2.1",
     "grunt-gitinfo": "0.1.6",
     "grunt-modernizr": "0.6.0",
     "grunt-s3": "0.2.0-alpha.3",
     "grunt-sass": "0.14.0",
-    "time-grunt": "0.3.2",
     "grunt-userevvd": "0.1.4",
+    "grunt-webpack": "git+ssh://git@github.com:dtothefp/grunt-webpack.git",
+    "handlebars-loader": "0.1.7",
+    "imports-loader": "0.6.3",
     "load-grunt-config": "0.16.0",
     "optimizely-marked": "0.0.4",
-    "uglify-js": "2.4.15"
+    "script-loader": "0.6.1",
+    "time-grunt": "0.3.2",
+    "uglify-js": "2.4.15",
+    "webpack": "1.4.15",
+    "grunt-concurrent": "^1.0.0",
+    "jshint": "^2.6.0",
+    "jshint-loader": "^0.8.1"
   },
   "devDependencies": {
     "body-parser": "1.8.3",
@@ -52,17 +61,7 @@
     "grunt-resemble-cli": "0.0.8",
     "grunt-text-replace": "0.3.12",
     "nightmare": "git+ssh://git@github.com:dtothefp/nightmare.git",
-    "phantomjs": "~1.9.12",
-    "webpack": "^1.4.15",
-    "grunt-webpack": "git+ssh://git@github.com:dtothefp/grunt-webpack.git",
-    "banner-footer-webpack-plugin": "git+ssh://git@github.com:dtothefp/banner-footer-webpack-plugin.git",
-    "script-loader": "^0.6.1",
-    "handlebars": "^1.3.0",
-    "handlebars-loader": "^0.1.7",
-    "6to5": "^2.13.3",
-    "6to5-loader": "^2.0.0",
-    "exports-loader": "^0.6.2",
-    "imports-loader": "^0.6.3"
+    "phantomjs": "~1.9.12"
   },
   "licenses": [
     {
